server/requirements.txt
django
djangorestframework
jsonfield

# httpie
